refactor: Make lxd_configurer private module
